Restructure folder hierarchy
