+++ weartsdk/WeArtTrackingRawData.py
@@ -1,6 +1,6 @@
-from WeArtMessageListener import WeArtMessageListener
-from WeArtCommon import HandSide, ActuationPoint, SensorData
-from WeArtMessages import WeArtMessage, RawSensorsData
+from .WeArtMessageListener import WeArtMessageListener
+from .WeArtCommon import HandSide, ActuationPoint, SensorData
+from .WeArtMessages import WeArtMessage, RawSensorsData
 
 from dataclasses import dataclass
